% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{AlphasDeltas2AD}
\alias{AlphasDeltas2AD}
\title{Convert  the alpha's and delta's into A1, A2 and D.
See Xie & Xu (1999) and Doligez et al (2013).
By definition:
A1 = (alphaA - alphaB) / 2
A2 = (alphaC - alphaD) / 2
We fix alphaA = alphaC = 0 as constraints.
Thus: 
alphaB = - 2 * A1 
alphaD = - 2 * A2
Also by definition:
D = deltaAC - deltaAD - deltaBC + deltaBD}
\usage{
AlphasDeltas2AD(seg, alphas, deltas)
}
\arguments{
\item{seg}{segregation type}

\item{alphas}{named vector}

\item{deltas}{named vector}
}
\value{
a vector of additive effect male and female and dominance effect
}
\description{
Convert  the alpha's and delta's into A1, A2 and D.
See Xie & Xu (1999) and Doligez et al (2013).
By definition:
A1 = (alphaA - alphaB) / 2
A2 = (alphaC - alphaD) / 2
We fix alphaA = alphaC = 0 as constraints.
Thus: 
alphaB = - 2 * A1 
alphaD = - 2 * A2
Also by definition:
D = deltaAC - deltaAD - deltaBC + deltaBD
}
