% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicfl2.R
\name{bic.fl2}
\alias{bic.fl2}
\title{Description of the function}
\usage{
bic.fl2(response, regressors, group, seq.a = NULL, lambda = NULL,
  mina = 0.1, plot = TRUE)
}
\arguments{
\item{response}{a vector response variable}

\item{regressors}{a quantitative matrix of regressor}

\item{group}{a vector with two levels. (The group of the ANCOVA)}

\item{lambda}{if the user wants to use it owns values of lambdas}

\item{a}{the parameters that indicate how much the coefficients will be fused}
}
\value{
The coefficients of the fused lasso ANCOVA for the different value of lambda
}
\description{
Description of the function
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- bic.fl2(y, regressors, group)
print(mod[[2]])
}
