% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{CalcBeta}
\alias{CalcBeta}
\title{CalcBeta}
\usage{
CalcBeta(seg, marker, X, config, Alpha = NULL, Delta = NULL, sigma2A,
  sigma2D, VbA, VbD)
}
\arguments{
\item{seg}{segregation type}

\item{marker}{marker name with an effect on phenotype}

\item{X}{design matrix of genotypic markers}

\item{config}{2 dimension vector which indicates the presence (1) or the absence (0) of a QTL}

\item{Alpha}{additive value of the allele, if NULL, genetic effects are drawn in a normal distrbution}

\item{Delta}{Dominance value, if NULL, genetic effects are drawn in a normal distrbutio}

\item{sigma2A}{additive genetic variance}

\item{sigma2D}{dominance genetic variance}

\item{VbA}{matrix of variance covariance of additive effect}

\item{VbD}{matrix of variance covariance of dominance effect}
}
\value{
a df Betas which contains 3 columns, B ; B.1 ; B.2, some of them are null according to the configuration
}
\description{
CalcBeta
}
