% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvfl2.R
\name{cv.fl2}
\alias{cv.fl2}
\title{Fuse lasso}
\usage{
cv.fl2(response, regressors, group, seq.a = NULL, lambda = NULL,
  mina = 0.1, nfold = 5, nrep = 10, nb.cores = 3, plot = TRUE)
}
\arguments{
\item{response}{a vector response variable}

\item{regressors}{a quantitative matrix of regressor}

\item{group}{a vector with two levels. (The group of the ANCOVA)}

\item{seq.a}{the parameters that indicate how much the coefficients will be fused. (The more a is small the more the coefficient will be supposed to fused). If NULL the sequence to test will be calculted automaticaly}

\item{lambda}{if the user wants to use it owns values of lambdas}

\item{mina}{if seq.a is NULL the minimum value of a that is use in seq.a ( a must be higher than 0.)}

\item{nrep}{the number of times we perform each CV for the different value of a in seq.a}
}
\value{
The ddl and the error of the lambda that minimise the CV error for differents value of a (we take the mean over nrep  replicats)
}
\description{
Cross-validation to find a (and optionally lambda).
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- cv.fl2(y, regressors, group)
}
\seealso{
\code{\link{cv_fl2_fixa}}
}
