% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{fus2modQTL}
\alias{fus2modQTL}
\title{fus2modQTL}
\usage{
fus2modQTL(X, Y, marker, map, group = c("R1", "R2"), nrep = c(5, 100),
  plot = c(FALSE, TRUE), cutoff = 0.85, lambda = NULL, nb.cores)
}
\arguments{
\item{X}{design matrix for genotypic markers}

\item{Y}{data frame of response (multivariate)}

\item{marker}{marker which has an effect on phenotype}

\item{map}{genetic map to localize markers}

\item{group}{name of 2 group responses, by default it is "R1" and "R2"}

\item{nrep}{2 dimension vector for number repetition for cross validation and stability selection}

\item{plot}{logical, plot or not of stability path}

\item{cutoff}{after which probability do we select predictors after stability selection}

\item{lambda}{value of lambda for lasso parameter, default is null}

\item{nb.cores}{result (-1) of the function DetectCores}
}
\description{
fus2modQTL
}
