% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{get_group}
\alias{get_group}
\title{Description of the function
This function indicate the group among the collumn of a matrix.}
\usage{
get_group(X, sep = "\\\\.")
}
\arguments{
\item{X}{a matrix with colnames indicating group variables}

\item{sep}{the separators between the group (for exemlpe the marker) and the specificity of the columns (for exemple the allele)}
}
\description{
Description of the function
This function indicate the group among the collumn of a matrix.
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
}
