% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{gglasso_st_tot}
\alias{gglasso_st_tot}
\title{Description of the function}
\usage{
gglasso_st_tot(X, Y, group = NULL, sep = "\\\\.", nb.cores = 7, B = 500,
  PFER = 1, type_group, verbose = FALSE)
}
\description{
Description of the function
}
