% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{glmnetmultivarQTL}
\alias{glmnetmultivarQTL}
\title{glmnetmultivarQTL}
\usage{
glmnetmultivarQTL(X, Y, marker, map, choose.alpha = 0.1, PFER = 1,
  B = 500, cutoff = 0.85, plot = TRUE, nb.cores)
}
\arguments{
\item{X}{design matrix of genotypic markers}

\item{Y}{data frame of response}

\item{marker}{marker name (character) which has an effect}

\item{map}{genetic map, 2 columns V1 and V2, one for the marker and the other for the position}

\item{choose.alpha}{value for parameter alpha in elastic net}

\item{PFER}{per family error rate for stability selection}

\item{B}{number of subsampling to apply for stability selection}

\item{cutoff}{after which probability do we select predictors after stability selection}

\item{plot}{logical, plot or not of stability paths}

\item{nb.cores}{result (-1) of the function DetectCores}
}
\description{
glmnetmultivarQTL
}
