% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{grp_lasso}
\alias{grp_lasso}
\title{Description of the function}
\usage{
grp_lasso(X, Y, group)
}
\arguments{
\item{X}{a matrix with different Marker and different allele}

\item{Y}{a matrix or a vector with one or several Trait}

\item{group}{a vector indicating which columns are togetrher ( of length (number of columns of Y * number of columns of X))}
}
\value{
The coefficients of the fused lasso ANCOVA for the different value of lambda
}
\description{
Description of the function
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
}
