% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{MeanResponseOnGenoclass}
\alias{MeanResponseOnGenoclass}
\title{For each allele i, compute its additive effect (alpha_i), where i is in {a,b,c,d} if segregation is abxcd, etc.
Moreover, for each parent (par1 of genotype ab and par2 of genotype cd), 
compute half of its "average effect of allelic substitution" (alpha).
Ref: Lynch & Walsh (1998) and Doligez et al (2013)
If <abxcd>:
Af = half.alpha.par1 = (alphaA - alphaB) / 2 = ((muAD + muAC) - (muBD + muBC)) / 4
Am = half.alpha.par2 = (alphaC - alphaD) / 2 = ((muAC + muBC) - (muAD + muBD)) / 4
Indeed, muAD = alphaA + alphaD
If <efxeg>:
Af = (alphaE - alphaF) / 2 = ((muEG + muEE) - (muFG + muFE)) / 4
Am = (alphaE - alphaG) / 2 = ((muEE + muFE) - (muEG + muFG)) / 4
If <hkxhk>:
Af = Am = (alphaH - alphaK) / 2 = (muHH - muKK) / 4
If <lmxll>:
Af = (alphaL _ alphaM) / 2 = (muLL - muLM) / 2
Am = 0, no effect of allelic substitution for homozygous parents
If <nnxnp>:
Af = 0
Am = (alphaN - alphaP) / 2 = (muNN - muNP) / 2}
\usage{
MeanResponseOnGenoclass(clgeno, marker, response, cat.response,
  NA.value = "--")
}
\arguments{
\item{clgeno}{data frame, individuals are in column (colnames = geno), markers are in row (rownames = marker name)}

\item{marker}{a specific rowname from clgeno}

\item{response}{a data frame containing the response and the individuals (it must be in a column)}

\item{cat.response}{category of response (use as ylab in the plot)}

\item{NA.value}{how are coded NA in the clgeno data frame ? Defaut "--"}
}
\value{
data frame with a column Am for the male additive effect and Af for the female additive effect
}
\description{
For each allele i, compute its additive effect (alpha_i), where i is in {a,b,c,d} if segregation is abxcd, etc.
Moreover, for each parent (par1 of genotype ab and par2 of genotype cd), 
compute half of its "average effect of allelic substitution" (alpha).
Ref: Lynch & Walsh (1998) and Doligez et al (2013)
If <abxcd>:
Af = half.alpha.par1 = (alphaA - alphaB) / 2 = ((muAD + muAC) - (muBD + muBC)) / 4
Am = half.alpha.par2 = (alphaC - alphaD) / 2 = ((muAC + muBC) - (muAD + muBD)) / 4
Indeed, muAD = alphaA + alphaD
If <efxeg>:
Af = (alphaE - alphaF) / 2 = ((muEG + muEE) - (muFG + muFE)) / 4
Am = (alphaE - alphaG) / 2 = ((muEE + muFE) - (muEG + muFG)) / 4
If <hkxhk>:
Af = Am = (alphaH - alphaK) / 2 = (muHH - muKK) / 4
If <lmxll>:
Af = (alphaL _ alphaM) / 2 = (muLL - muLM) / 2
Am = 0, no effect of allelic substitution for homozygous parents
If <nnxnp>:
Af = 0
Am = (alphaN - alphaP) / 2 = (muNN - muNP) / 2
}
\seealso{
AD2AlphaDelta, AlphasDeltas2AD
Af = half.alpha.par1 = ((muAD + muAC) - (muBD + muBC)) / 4
Am = half.alpha.par2 = ((muAC + muBC) - (muAD + muBD)) / 4
Af = ((muEG + muEE) - (muFG + muFE)) / 4
Am = ((muEE + muFE) - (muEG + muFG)) / 4
Af = Am = (muHH - muKK) / 4
Af = (muLL - muLM) / 2
Am = 0
Af = 0
Am = (muNN - muNP) / 2
}
