% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{SimulPhenosBiparentalCross}
\alias{SimulPhenosBiparentalCross}
\title{Phenotypes are simulated for offsprings from a bi-parental cross between outbred parents (AB x CD):
y = g + e where y is the phenotyic value, g the genotypic value and e the error.
See Xie & Xu (1999) and Doligez et al (2013).
By definition:
For a marker segregating <abxcd> :
A1 = (alphaA - alphaB) / 2
A2 = (alphaC - alphaD) / 2
We fix alphaA = alphaC = 0 as constraints.
Thus: 
alphaB = - 2 * A1 
alphaD = - 2 * A2
Also by definition:
D = deltaAC - deltaAD - deltaBC + deltaBD
We also fix DeltaAD = DeltaBC = 0 (see also param constraints.deltas).}
\usage{
SimulPhenosBiparentalCross(clgeno, marker, Gij, A1, A2, D,
  constraints.deltas = c(0.5, 0.5), h2, plot.type = "beanplot",
  ylim = NULL, logy = "")
}
\arguments{
\item{clgeno}{data frame of marker genotypes with individuals in columns (colnames = geno, first column contains the marker segregation), markers in rows (rownames = marker name)}

\item{Gij}{mean response value on genotypic class,  given in the funtion MeanResponseOnGenoclass}

\item{A1}{half of the additive effect of the parent 1, given in the funtion MeanResponseOnGenoclass}

\item{A2}{half of the additive effect of the parent 2, given in the funtion MeanResponseOnGenoclass}

\item{D}{(half of?) the dominance effect}

\item{constraints.deltas}{by default DeltaAC = DeltaBD = D / 2, 
vector of weights for each Delta, the sum is equal to 1}

\item{h2}{heritability}

\item{plot.type}{whether "beanplot" or "boxplot"}

\item{ylim}{default = NULL, limits of the y axis}

\item{qtl}{a specific rowname from clgeno which will be the QTL (only one for the moment)}
}
\value{
a list of several elements :
- Delta : the dominance effect
- Alpha : half the additive effect of each parent
- data : a data frame containing the phenotype in column pheno, 
the genotypic effect in column g, 
the error term in column e,
the genotypic class at the QTL in column geno.qtl
}
\description{
For a marker segregating as <efxeg>: (A=E B=F C=E D=G)
alphaE = 0
alphaF = - 2 * A1
alphaG = - 2 * A2
deltaEG = deltaEF = 0
deltaFG = D
}
\details{
For the cross <hkxhk> :
AlphaH = 0
AlphaK = - 2 * A1
DeltaHH = 0
DeltaHK = D / 2
DeltaKK = D / 2

For the cross <lmxll>
AlphaL = 0
AlphaM = - 2 * A1 
DeltaLM = 0

For the cross <nnxnp>
AlphaN = 0
AlphaP = -2 * A1
DeltaNP = 0

Additive genetic variance according to Xie & Xu :
sigma2A = A1^2 +A2^2
and also dominance genetic variance :
sigma2D = (1/16) * D^2
}
