% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{SimulResponse}
\alias{SimulResponse}
\title{This function is aiming to simulate genotypic values and response(s) 
in a segregating population where only 1 marker has an effect. 
In a multivariate case, the multiple linear regression besomes :
    Y = XB +  E 
Y is a matrix of NxT dimension with N observations and T responses}
\usage{
SimulResponse(marker, X, seg, config = c(0), h2 = c(0),
  mean.response = c(0), rhoE = 0, Alpha = NULL, Delta = NULL,
  sigma2A = c(0.1), sigma2D = c(0.1), rhoB = 0, sigma2.E = 0)
}
\arguments{
\item{marker}{which marker has an impact on the phenotype}

\item{X}{design matrix additively coded in 0,1,2; the number of individuals
genotypes should be n. Rownames(X) should correspond to the marker names.}

\item{seg}{is the segregation observed at the marker, written in this format : "<mmxmm>" with m letters
according to the updated Joinmap format, for example for a 4 alleles segregation : "<abxcd>"}

\item{config}{a 2 dimension vector, if c(1), assume that there is one condition with a QTL,
if config = c(1,0), there is a QTL in the condition 1 but not 2, inverse for c(0,1),
c(0,0) = no QTL and c(1,1) = QTL in both conditions.
The dimension of config indicates the dimension of the response : 1 or 2 columns}

\item{h2}{heritability (0<h2<1), amount of noise to add to simulate phenotypes based on real genetic effect: if h2 = 1, there is no noise.
Dimension of h2 = dimension of the response (and of config). If there is no QTL, h2 = 0}

\item{mean.response}{of the response; same dimension as response (1 or 2 columns). By default, mean.response = 0}

\item{rhoE}{correlation between the 2 conditions (not due to genetic variation) in the variance of the error term}

\item{Alpha}{half of the additive effect of each parent, calculated in MeanResponseOnGenoClass function
If Alpha = NULL, the beta's are drawn from an univariate or multivariate normal distribution with a mean of 0
and a standard deviation of sigma2B}

\item{Delta}{dominance effect, calculated in MeanResponseOnGenoClass function.
If Alpha = NULL, Delta must be NULL too}

\item{sigma2A}{additive genetic variance (in the statistical model) in the condition where there is a QTL, same dimension as response. 
If there is no QTL, sigma2A must be equal to 0}

\item{sigma2D}{dominance genetic variance (in the statistical model), must be equal to 0 if there is no QTL, 
can be 0 in the response where there  is a QTL}

\item{rhoB}{correlation between the 2 conditions in the variance of the B term, only used if config = c(1,1)}

\item{sigma2.E}{is the error variance. In config [11] or [1], the error variance is defined by h2 and sigma2G only. If 
config is [01] or [10], one error variance is not defined, we have to add it separately, it is sigma2.E.}
}
\value{
list of 3 elements : 
- y which is the response matrix (phenotype), with 1 or 2 columns according to config
- B which is the marker effect, with 1 or 2 columns according to config
- X the genotypic matrix
}
\description{
X is the design matrix of markers (NxP) with P predictors, here markers
}
\details{
B is the marker effects matrix. 2 choices: 
   - B is calculated from Alphas and Deltas from the previous
function 
   - B is drawn from a Matrix Normal distribution : B ~ MN(0,UbVb)
      We assume that Ub = Id (no covariance between markers)
      Vb is a TxT variance covariance matrix between responses
      VbA is the matrix with additive genetic effect based on the statistical model and
      VbD is the matrix with dominance statistical effect
      VbA =  | sigma²1A             rhoB.sigma1A.sigma2A |
            | rhoB.sigma1A.sigma2A   sigma²2A           |
            rhoB is the covariance between the responses due to genetics
            
E is the matrix of residual error (NxT)
   E ~ MN(0,UeVe)
   - we assume Ue = Id : no covariance between individuals due to error because we have 
   a good experimental design
   - Ve =  | sigma²1e             rhoE.sigma1e.sigma2e |
           | rhoE.sigma1e.sigma2e   sigma²2 e          |
           With rhoE the correlation between responses not due to genetic variation
Sigma²1e and Sigma²2e are calculated from Xie & Xu formula with A1, A2 and D
}
