% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4GenoClass.R
\name{univQTL}
\alias{univQTL}
\title{univQTL}
\usage{
univQTL(dat.doligez2013, response, nperm = 100, qtl.dist = 5,
  method = "em", plot = TRUE, ylim = c(0, 15), main)
}
\arguments{
\item{dat.doligez2013}{a 4way cross class object}

\item{response}{a data frame with 1 column which contains the response to analyze, rownames must
be individual names}

\item{nperm}{number of permutation to apply to find the LOD threshold, by default 100}

\item{qtl.dist}{confidence interval to apply on both sides of the marker}

\item{method}{by default "em"}

\item{plot}{logical, to plot or not the result of the scanone function}

\item{ylim}{limit of the plot}
}
\value{
logical, TRUE if the QTL was above the threshold, FALSE else
}
\description{
univQTL
}
